Improved support almost-compatible point structures
In FontMath and Superpolator it is accepted that straight sections can
interpolate with curves by adding on-point controls to all straight
lines. This causes some extra effort calculations but makes it easier
to make complex shapes. In general use these mathShapes seem plenty
fast and I think it is a allowed to add the extra points. Some tests
are added. Version number upped.