Experimental three-by-one designspace
Narrow master to square master to wide master. The "medium-thin" glyph
is a point compatible intermediate. Size factors are recalculated to
fit the model.
Discussion: While the square appears to be the most logical breakpoint
from a geometric point of view, it is not at all certain that all
interpolations would have a square. Perhaps it would be more useful to
set intermediates at arbitrary breakpoints. Of course that raises the
question if those masters need to be point compatible. And then the
whole thing blows up.